Delete unused TextureFormat and update imports
Removed the `TextureFormat.kt` file due to its redundancy. Updated the package imports across multiple files from `webgpu` to `webgpu.native` for consistency and to reflect the updated package structure.

+++ wgpu4k/src/commonMain/kotlin/io.ygdrasil.wgpu/Buffer.kt
@@ -8,17 +8,17 @@ expect class Buffer : AutoCloseable {
 
     fun unmap()
 
-    fun mapFrom(buffer: ShortArray, offset: Int = 0)
+    fun mapFrom(buffer: ShortArray, offset: ULong = 0u)
 
-    fun mapFrom(buffer: FloatArray, offset: Int = 0)
+    fun mapFrom(buffer: FloatArray, offset: ULong = 0u)
 
-    fun mapFrom(buffer: ByteArray, offset: Int = 0)
+    fun mapFrom(buffer: ByteArray, offset: ULong = 0u)
 
-    fun mapInto(buffer: ByteArray, offset: Int = 0)
+    fun mapInto(buffer: ByteArray, offset: ULong = 0u)
 
-    fun mapInto(buffer: IntArray, offset: Int = 0)
+    fun mapInto(buffer: IntArray, offset: ULong = 0u)
 
-    suspend fun map(mode: Set<MapMode>, offset: GPUSize64 = 0, size: GPUSize64 = this.size)
+    suspend fun map(mode: Set<MapMode>, offset: GPUSize64 = 0u, size: GPUSize64 = this.size)
 
     override fun close()
 }
