+++ wgpu4k/src/commonMain/kotlin/io.ygdrasil.wgpu/Types.kt
@@ -8,15 +8,15 @@ typealias GPUColorWriteFlags = Int
 
 typealias GPUDepthBias = Int
 
-typealias GPUFlagsConstant = Int
+typealias GPUFlagsConstant = ULong
 
-typealias GPUIndex32 = Int
+typealias GPUIndex32 = UInt
 
-typealias GPUIntegerCoordinate = Int
+typealias GPUIntegerCoordinate = UInt
 
 typealias GPUIntegerCoordinates = Pair<GPUIntegerCoordinate, GPUIntegerCoordinate>
 
-typealias GPUIntegerCoordinateOut = Int
+typealias GPUIntegerCoordinateOut = UInt
 
 typealias GPUMapModeFlags = Int
 
@@ -28,11 +28,11 @@ typealias GPUShaderStageFlags = Int
 
 typealias GPUSignedOffset32 = Int
 
-typealias GPUSize32 = Int
+typealias GPUSize32 = UInt
 
-typealias GPUSize32Out = Int
+typealias GPUSize32Out = UInt
 
-typealias GPUSize64 = Long
+typealias GPUSize64 = ULong
 
 typealias GPUSize64Out = Long
 
@@ -41,21 +41,21 @@ typealias GPUStencilValue = Long
 typealias GPUTextureUsageFlags = Int
 
 data class Size3D(
-	val width: Int,
-	val height: Int = 1,
-	val depthOrArrayLayers: Int = 1,
+	val width: UInt,
+	val height: UInt = 1u,
+	val depthOrArrayLayers: UInt = 1u,
 ) {
     fun toArray() = arrayOf(width, height, depthOrArrayLayers)
 }
 
 data class Origin3D(
-	val x: GPUIntegerCoordinate = 0,
-	val y: GPUIntegerCoordinate = 0,
-	val z: GPUIntegerCoordinate = 0,
+	val x: GPUIntegerCoordinate = 0u,
+	val y: GPUIntegerCoordinate = 0u,
+	val z: GPUIntegerCoordinate = 0u,
 ) {
     fun toArray() = arrayOf(x, y, z)
 
     companion object {
-        val Zero = Origin3D(0, 0, 0)
+        val Zero = Origin3D(0u, 0u, 0u)
     }
 }
